% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eff}
\alias{eff}
\title{Efficacy response generation parameters}
\format{An object of class \code{list} of length 4.}
\usage{
eff
}
\value{
\item{Dose_Cycle_Meff}{Dose-cycle mean efficacy matrix. A 4 dimension array
providing the mean of the multivariate Gaussian distribution in
efficacy data generation. The dimension of the Dose-cycle mean efficacy matrix
is 5 5 6 6 which represents dose efficacy pattern, cycle efficacy pattern,
dose and cycle. Patterns are ordered from increasing, flat, platform decreasing
and quadratic efficacy across dose levels and cycles.}

\item{Sigma}{A 6 by 6 matrix, the covariance matrix of the multivariate
Guassian distribution in efficacy data generation.}

\item{sd_trans}{A positive number controls the skewness of the distribution
of the efficacy response}

\item{eff.M}{An array recording the mean of the generated efficacy data.
The dimension of the Dose-cycle mean efficacy matrix
is 5 5 6 6 which represents dose efficacy pattern, cycle efficacy pattern,
dose and cycle. Patterns are ordered from increasing, flat, platform decreasing
and quadratic efficacy across dose levels and cycles.}
}
\description{
A list of 3 for the parameters in the efficacy response generation
}
\keyword{datasets}
