% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{patlist.display}
\alias{patlist.display}
\title{Display a patient records}
\usage{
patlist.display(patlist, n.dose, n.cycle)
}
\arguments{
\item{n.cycle}{}
}
\description{
Display a patient records
}
