% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patlist_sim}
\alias{patlist_sim}
\title{A list of patient information}
\format{An object of class \code{list} of length 6.}
\usage{
patlist_sim
}
\value{
\item{PatID}{denotes the patient ID where the elements are specified by
cohort and subject number. For example, "cohort2subject3" denotes the third
subject in the second cohort}
\item{dose}{records the dose level assigned for each patient
through the whole treatment}
\item{cycle}{shows the treatment cycle information of each record}
\item{nTTP}{records th corresponding nTTP score.}
\item{dlt}{indicates whether a DLT event is observed or not?}
\item{efficacy}{provides the continuous efficacy for each cycle. Required when
\code{effcy.flag == T}. The range of efficacy is (0, 1) and use -1 for
missing efficacy response.}
}
\description{
A list of the patient treatment records. The simulation with MTD = 4,
dose-toxicity trend, efficacy-toxicity trend and efficacy-cycle trend are all
flat.
}
\keyword{datasets}
