% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1model.R
\name{phase1model}
\alias{phase1model}
\title{Simulation for an Adaptive, Multi-Stage Phase I Dose-Finding Design}
\usage{
phase1model(seed = 1234, patlist = NULL, patID_act = NULL,
  cycle_act = NULL, dose_act = NULL, dlt_act = NULL, doses = 1:6,
  cycles = 1:6, tox.target = 0.28, p_tox1 = 0.2, p_tox2 = 0.2,
  trialSize = 36, chSize = 3, thrd1 = 0.28, thrd2 = 0.28,
  proxy.thrd = 0.1, param.ctrl = list(), inits.list.set = list(),
  n.chains = 1, effcy.flag = T, ICD.flag = T, DLT.drop.flag = T,
  testedD = T, IED.flag = T, ICD_thrd = 0.3)
}
\arguments{
\item{seed}{The seed of R's random number generator. Default is 1234}

\item{patlist}{A list of the patient treatment records,
which must contains the following variables:
\describe{
\item{PatID}{denotes the patient ID where the elements are specified by
cohort and subject number. For example, "cohort2subject3" denotes the third
subject in the second cohort}
\item{dose}{records the dose level assigned for each patient
through the whole treatment}
\item{cycle}{shows the treatment cycle information of each record}
\item{nTTP}{records the corresponding nTTP score.}
\item{dlt}{indicates whether a DLT event is observed or not?}
\item{efficacy}{provides the continuous efficacy for each cycle. Required when
\code{effcy.flag == T}. The range of efficacy is (0, 1) and use -1 for
missing efficacy response.}
}}

\item{patID_act}{A vector records the patients' ID who need dose
recommendation for next cycle}

\item{cycle_act}{A vector records the current cycle of patID_act.}

\item{dose_act}{A vector records the current dose level of patID_act}

\item{dlt_act}{A vector indicates whether a dlt is observed in current cycle for current patients}

\item{doses}{A vector of doses that users are going to explore.
Default is 1:6, where dose 1 through dose 6 are being tested.}

\item{cycles}{A vector of cycles that the treatment plans to go through.
Default is 1:6, where patients will experience up to 6 cycles of the treatment}

\item{tox.target}{The target toxicity of the treatment. Default is 0.28.
See details below.}

\item{p_tox1}{The probability cutoff for cycle 1 toxicity. Default is 0.2.
See details below.}

\item{p_tox2}{The probability cutoff for later cycles toxicity beyond cycle 1.
Default is 0.2. See Details below.}

\item{trialSize}{The maximum sample size for trial simulation. Default is 36.
Must be the multiple of cohort size, represented by chSize}

\item{chSize}{The cohort size of patients recruited. Default is 3.}

\item{thrd1}{An upper bound of toxicity for cycle 1 of the treatment.
Default is 0.28. See Details below.}

\item{thrd2}{An upper bound of toxicity for late cycles of the treatment,
beyond cycle 1. Default is 0.28. See Details below}

\item{proxy.thrd}{A distance parameter to define efficacious doses.
Any dose whose predicted efficacy is within proxy.thrd away from the
largest one among the safe doses will be declared an efficacious dose.}

\item{param.ctrl}{A list specifying the prior distribution for the parameters.
p1_beta_intercept, the prior mean of intercept of toxicity model assuming
a normal prior; p2_beta_intercept, the precision (inverse of variance) of
intercept of toxicity model assuming a normal prior; p1_beta_cycle,
the prior mean of cycle effect of toxicity model assuming a normal prior;
p2_beta_cycle, the precision (inverse of variance) of cycle effect of
toxicity model assuming a normal prior; p1_beta_dose, the prior minimum of
dose effect of toxicity model assuming a uniform prior; p2_beta_dose,
the prior maximum of dose effect of toxicity model assuming a uniform
prior; p1_alpha, the prior mean vector of the parameters from efficacy
model assuming a multivariate normal prior; p2_alpha, the prior precision
 matrix (inverse of covariance matrix) of the parameters from efficacy
 model assuming a multivariate normal prior; p1_gamma0, the prior mean
 of association parameter \eqn{\gamma} (See Du et al(2017)) of two submodels of the
 joint model assuming a normal prior; p2_gamma0, the prior precision
 (inverse of variance) of association parameter \eqn{\gamma} of two submodels of the
 joint model assuming a normal prior. Default is non-informative priors.}

\item{inits.list.set}{Default is list()}

\item{n.chains}{No. of MCMC chains in Bayesian model fitting. Default is 1}

\item{effcy.flag}{Whether efficacy data is considered in the model fitting or not}

\item{ICD.flag}{Whether we allow dose changing for cycle > 1 in stage 1
model or not? Default is TRUE. See details below}

\item{DLT.drop.flag}{Whether the patients should suspend the treatment
when observing DLT. Default is TRUE}

\item{testedD}{Default is TRUE. Whether we only allow ICD or IED among
cycle 1 tested dose level}

\item{IED.flag}{Default is TRUE. Whether we allow dose changing for
cycle > 1 in stage 2 model or not?}

\item{ICD_thrd}{The cut-off point of the posterior toxicity probability in
defining ICD. Default is 0.3. See details below.}
}
\value{
\item{doseA}{The recommended dose level for cycle 1 for new cohorts}
\item{pat_rec}{The recommended dose for current patients for next cycle}
}
\description{
A function to implement simulation for an adaptive, multi-stage phase 1
dose-finding design incorporating a longitudinal continuous efficacy outcome
and toxicity data from multiple treatment cycles, proposed by Du et al(2017)
and grade
}
\examples{

data(patlist)
patlist <- patlist_sim[-((length(patlist_sim$PatID) - 2):
                       length(patlist_sim$PatID))]
patID_act <- patlist_sim$PatID[(length(patlist_sim$PatID) - 2):
                               length(patlist_sim$PatID)]
cycle_act <- patlist_sim$cycle[(length(patlist_sim$PatID) - 2):
                                length(patlist_sim$PatID)] - 1
dose_act <- rep(1, 3)
dlt_act <- rep(0, 3)

test <- phase1model(patlist = patlist, patID_act = patID_act,
                    cycle_act = cycle_act, dose_act = dose_act,
                    dlt_act = dlt_act, trialSize = 36, chSize = 3)

test$doseA
test$pat_rec

test2 <- phase1model(patlist = patlist, patID_act = patID_act,
                    cycle_act = cycle_act, dose_act = dose_act,
                    dlt_act = dlt_act, trialSize = 75, chSize = 3)
test2$doseA
test2$pat_rec

\dontrun{

}

}
