% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhaseIPRMD.R
\docType{package}
\name{Phase1PRMD}
\alias{Phase1PRMD}
\alias{Phase1PRMD-package}
\title{Phase1PRMD: A package for model-based phase I designs with both clinician-assessed
AEs and PROs as toxicity endpoints, to improve accuracy in dose selection and
to incorporate individualized dosing based on patient’s tolerance from prior
treatment cycles.}
\description{
The PhaseI package provides three categories of important functions:
foo, bar and baz.
}
\section{scenario summary functions}{

The foo functions ...
}

