% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{plot.RunPRMD}
\alias{plot.RunPRMD}
\title{nTTP and efficacy boxplots of a RunPRMD object}
\usage{
\method{plot}{RunPRMD}(x, ..., select_cycle = x$cycles)
}
\arguments{
\item{x}{RunPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}

\item{select_cycle}{A vector indication the cycle in the boxplot.
Default is \code{cycle} of \code{x}.}
}
\description{
nTTP and efficacy boxplots of a RunPRMD object
}
\examples{
## Check ?RunPRMD for example

}
