% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SimPRMD.R
\name{plot.SimPRMD}
\alias{plot.SimPRMD}
\title{Plots of a SimPRMD object}
\usage{
\method{plot}{SimPRMD}(x, ..., title.add = TRUE)
}
\arguments{
\item{x}{SimPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}

\item{title.add}{controls whether there is a title on plots or not.}
}
\description{
Plots of a SimPRMD object.
}
\examples{
## Check ?SimPRMD for example

}
