% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prob}
\alias{prob}
\title{Toxicity probability matrix}
\format{An object of class \code{array} of dimension 4 x 3 x 6 x 6 x 3 x 5.}
\usage{
prob
}
\description{
A 6 dimension array providing the toxicity probability.
The dimension of the toxicity probability matrix is 4 3 6 6 3 5 which
represents scenario, cycle effect, dose level, cycle number, tox type,
and tox grade. Scenarios are order in a way that first scenario is MTD = dose 2,
and second is MTD = dose 3, and third is MTD = dose 4, and fourth
is MTD = dose level 5. Cycle effect is ordered from decreasing toxicity
over cycles, flat, and increasing.
}
\references{
???
}
\keyword{datasets}
