% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phaseI.R
\name{sim_phase1}
\alias{sim_phase1}
\title{Simulation for an Adaptive, Multi-Stage Phase I Dose-Finding Design}
\usage{
sim_phase1(seed = 1234, numTrials = 100, doses = 1:6, cycles = 1:6,
  eff.structure = matrix(0, nrow = 6, ncol = 6), eff.Sigma = diag(6),
  eff.sd_trans = 1, tox.target = 0.28, p_tox1 = 0.2, p_tox2 = 0.2,
  trialSize = 36, chSize = 3, thrd1 = 0.28, thrd2 = 0.28,
  proxy.thrd = 0.1, tox.matrix = NULL, wm = matrix(c(0, 0.5, 0.75, 1,
  1.5, 0, 0.5, 0.75, 1, 1.5, 0, 0, 0, 0.5, 1), byrow = T, ncol = 5),
  toxmax = 2.5, toxtype = NULL, intercept.alpha = NULL,
  coef.beta = NULL, cycle.gamma = NULL, param.ctrl = list(),
  inits.list.set = list(), n.chains = 1, effcy.flag = T,
  ICD.flag = T, DLT.drop.flag = T, testedD = T, IED.flag = T,
  ICD_thrd = 0.3)
}
\arguments{
\item{seed}{The seed of R's random number generator. Default is 1234}

\item{numTrials}{An integer specifying the number of simulations}

\item{doses}{A vector of doses that users are going to explore.
Default is 1:6, where dose 1 through dose 6 are being tested.}

\item{cycles}{A vector of cycles that the treatment plans to go through.
Default is 1:6, where patients will experience up to 6 cycles of the treatment}

\item{eff.structure}{A matrix provides the mean of the multivariate
Gaussian distribution in efficacy data generation. Specifically,
the \eqn{(i, j)}th element represents the mean value of \eqn{i}th dose level and
\eqn{j}th cycle of the Gaussian distribution for efficacy data generation.
Default is a \eqn{6 \times 6} zero matrix}

\item{eff.Sigma}{The covariance matrix of the multivariate Guassian
distribution in efficacy data generation.See details below.}

\item{eff.sd_trans}{A positive number controls the skewness of the distribution
of the efficacy response. Default is 1.5. See details below.}

\item{tox.target}{The target toxicity of the treatment. Default is 0.28.
See details below.}

\item{p_tox1}{The probability cutoff for cycle 1 toxicity. Default is 0.2.
See details below.}

\item{p_tox2}{The probability cutoff for later cycles toxicity beyond cycle 1.
Default is 0.2. See Details below.}

\item{trialSize}{The maximum sample size for trial simulation. Default is 36.
Must be the multiple of cohort size, represented by chSize}

\item{chSize}{The cohort size of patients recruited. Default is 3.}

\item{thrd1}{An upper bound of toxicity for cycle 1 of the treatment.
Default is 0.28. See Details below.}

\item{thrd2}{An upper bound of toxicity for late cycles of the treatment,
beyond cycle 1. Default is 0.28. See Details below}

\item{proxy.thrd}{A distance parameter to define efficacious doses.
Any dose whose predicted efficacy is within proxy.thrd away from the
largest one among the safe doses will be declared an efficacious dose.}

\item{tox.matrix}{Optional. A four-dimension array specifying the
probabilities of the occurrences of certain grades for certain types of
toxicities, at each dose level and cycle under consideration. Dimension 1
refers to doses; dimension 2 corresponds to cycles of the treatment;
dimension 3 regards the types of toxicities while dimenion 4 relates
to grades. If null, which is default choice, the arguments toxtype,
intercept.alpha, coef.beta, cycle.gamma must be provided to simulate
this array.}

\item{wm}{Clinical weight matrix, where toxicity types define the rows
while the toxicity grades define the columns. Usually solicited from physicians.}

\item{toxmax}{The normalization constant used in computing nTTP score.
For details, see Ezzalfani et al(2013).}

\item{toxtype}{Only specified when tox.matrix is null. This argument,
a character vector, specifies toxicity types considered in the trial.}

\item{intercept.alpha}{Only specified when tox.matrix is null.
A four element numeric vector specifying the intercepts for the
cumulative probabilities of the occurrences of grades 0-4 of
toxicities in proportional odds model. See Details below.}

\item{coef.beta}{Only specified when tox.matrix is null. A n numeric vector
specifying the slope for dose in proportional odds model for n types of toxicities. See Details below}

\item{cycle.gamma}{Only specified when tox.matrix is null. A scalar
controlling the cycle effect in simulation in proportional odds model.
See Details below}

\item{param.ctrl}{A list specifying the prior distribution for the parameters.
p1_beta_intercept, the prior mean of intercept of toxicity model assuming
a normal prior; p2_beta_intercept, the precision (inverse of variance) of
intercept of toxicity model assuming a normal prior; p1_beta_cycle,
the prior mean of cycle effect of toxicity model assuming a normal prior;
p2_beta_cycle, the precision (inverse of variance) of cycle effect of
toxicity model assuming a normal prior; p1_beta_dose, the prior minimum of
dose effect of toxicity model assuming a uniform prior; p2_beta_dose,
the prior maximum of dose effect of toxicity model assuming a uniform
prior; p1_alpha, the prior mean vector of the parameters from efficacy
model assuming a multivariate normal prior; p2_alpha, the prior precision
 matrix (inverse of covariance matrix) of the parameters from efficacy
 model assuming a multivariate normal prior; p1_gamma0, the prior mean
 of association parameter \eqn{\gamma} (See Du et al(2017)) of two submodels of the
 joint model assuming a normal prior; p2_gamma0, the prior precision
 (inverse of variance) of association parameter \eqn{\gamma} of two submodels of the
 joint model assuming a normal prior. Default is non-informative priors.}

\item{inits.list.set}{Default is list()}

\item{n.chains}{No. of MCMC chains in Bayesian model fitting. Default is 1}

\item{effcy.flag}{Whether the patients should suspend the treatment
when observing DLT. Default is TRUE}

\item{ICD.flag}{Whether we allow dose changing for cycle > 1 in stage 1
model or not? Default is TRUE. See details below}

\item{DLT.drop.flag}{Default is TRUE}

\item{testedD}{Default is TRUE. Whether we only allow ICD or IED among
cycle 1 tested dose level}

\item{IED.flag}{Default is TRUE. Whether we allow dose changing for
cycle > 1 in stage 2 model or not?}

\item{ICD_thrd}{The cut-off point of the posterior toxicity probability in
defining ICD. Default is 0.3. See details below.}
}
\value{
\item{senerio_sum}{contains \code{mnTTP.M} the matrix of mean nTTP for
each dose and cycle and \code{pDLT.M} matrix of probability of observing
DLT for each dose and cycle}
\item{alloc.perc}{The dose allocation percentage for cycle1}
\item{list_simul}{All simulations records}
\item{doses}{A vector of doses level of the study.}
\item{cycles}{A vector of cycles that the treatment plans to go through}
}
\description{
A function to implement simulation for an adaptive, multi-stage phase 1
dose-finding design incorporating a longitudinal continuous efficacy outcome
and toxicity data from multiple treatment cycles, proposed by Du et al(2017)
and grade
}
\examples{

data(prob)      # load prob.RData from package phaseI, Details see "?prob"
data(eff)       # load eff.RData from package phaseI. Details see "?eff"

eff.structure = eff$Dose_Cycle_Meff[2, 2, , ]
eff.Sigma = eff$Sigma
eff.sd_trans = eff$sd_trans

wm <- matrix(c(0, 0.5, 0.75, 1, 1.5,
               0, 0.5, 0.75, 1, 1.5,
               0, 0, 0, 0.5, 1),
             byrow = TRUE, ncol
              = 5)                          # weighted matrix for toxicity matrix
                                            # nrow = No.of type; ncol = No. of grade
toxmax <- 2.5
tox.matrix <- aperm(prob[3, 2, , , , ], c(3, 2, 1, 4)) # may want to change the dimension


#------- a flat dose-toxicity, dose-efficacy, cycle-efficacy pattern------#
#simul1 <- sim_phase1(numTrials = 1, tox.matrix = tox.matrix,
#                     eff.structure = eff.structure, eff.Sigma = eff.Sigma,
#                     eff.sd_trans = eff.sd_trans, wm = wm, toxmax = toxmax)

#------- a flat dose-toxicity pattern model ------#
simul2 <- sim_phase1(numTrials = 1, toxtype = c("H", "L", "M"),
                    intercept.alpha = c(1.9, 2.3, 2.6, 3.1),
                    coef.beta = c(-0.3, -0.2, -0.25),
                    cycle.gamma = 0, tox.target = 0.23,
                    thrd1 = 0.23, thrd2 = 0.23, p_tox1 = 0.2, p_tox2 = 0.2,
                    ICD.flag = FALSE, IED.flag = FALSE, effcy.flag = TRUE)

print(simul2)
summary(simul2)

\dontrun{

}

}
