% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{summary.RunPRMD}
\alias{summary.RunPRMD}
\title{Summary an RunPRMD object}
\usage{
\method{summary}{RunPRMD}(object, ...)
}
\arguments{
\item{object}{RunPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\value{
\item{object}{The output of function \code{RunPRMD}}
\item{mnttp.M}{The mean nTTP for all doses and cycles}
\item{dlt.count.M}{The number of DLT for all doses and cycles}
\item{eff.M}{The mean efficacy for all doses and cycles. Return \code{NULL}
when \code{object$effcy.flag == TRUE}}
}
\description{
Summary an RunPRMD object
}
\examples{
## Check ?RunPRMD for example

}
