% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sim_phase1.R
\name{summary.sim_phase1}
\alias{summary.sim_phase1}
\title{Displays a useful description of a sim_phase1 object}
\usage{
\method{summary}{sim_phase1}(object, ...)
}
\arguments{
\item{object}{sim_phase1 object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a useful description of a sim_phase1 object
}
\examples{
## Check ?sim_phase1 for example
}
