% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{patlist.display}
\alias{patlist.display}
\title{Display a patient records}
\usage{
patlist.display(patlist, n.dose, n.cycle)
}
\arguments{
\item{patlist}{The patient records}

\item{n.dose}{The number of dose in the study}

\item{n.cycle}{The number of cycle in the study}
}
\description{
Display a patient records
}
